requirements.txt
langchain
langchain-openai
langgraph
openai
tavily-python
playwright
beautifulsoup4
streamlit
python-dotenv
pydantic
requests
weasyprint
lxml
tqdm
reportlab